+++ srcs/main_app/front_end/urls.py
@@ -6,12 +6,6 @@ urlpatterns = [
     path('', views.index, name='index'),
     path('home/', views.homePage, name='home'),
     path('cards/', views.homeCards, name='home_cards'),
-    path('login/', views.loginPage, name='login'),
-    path('logout/', views.logout, name='login'),
     path('topbar/', views.topBar, name='topBar'),
-    path('token/', views.token, name='token'),
-    path('renew_token/', views.renew_token, name='renew_token'),
-    path('online/', views.onlineGame, name="onlineGame"),
-    path('offline/', views.offlineGame, name="offlineGame"),
     path('playerInfo/', views.getOpponentInfo, name='getOpponentInfo'),
 ]
