+++ srcs/main_app/game/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path('online/', views.onlineGame, name="onlineGame"),
+    path('offline/', views.offlineGame, name="offlineGame"),
+]
