+++ srcs/main_app/main_app/settings.py
@@ -45,7 +45,9 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'bootstrap5',
-    'front_end'
+    'front_end',
+    'login',
+    'game'
 ]
 
 MIDDLEWARE = [
